rename(ui-design-tool): rename PointerEventsPreventer to PointerEventsController
