+++ packages/ui-design-tool/components/InteractionController/InteractionController.tsx
@@ -19,7 +19,7 @@ import { useEvent, useEventListener } from '@pigyuma/react-utils';
 import React from 'react';
 import { AxisGrid } from '../AxisGrid/AxisGrid';
 import { HoveringOverlay } from '../HoveringOverlay/HoveringOverlay';
-import { PointerEventsPreventer } from '../PointerEventsPreventer/PointerEventsPreventer';
+import { PointerEventsController } from '../PointerEventsController/PointerEventsController';
 import { SelectionOverlay } from '../SelectionOverlay/SelectionOverlay';
 import * as styles from './InteractionController.css';
 import { InteractionControllerProps } from './types';
@@ -158,7 +158,7 @@ export const InteractionController: React.FC<InteractionControllerProps> = React
       <AxisGrid />
       <HoveringOverlay />
       <SelectionOverlay />
-      <PointerEventsPreventer />
+      <PointerEventsController />
     </div>
   );
 });
